Rename persons → person_names for consistency
The functions have always been `load_person_names` /
`load_person_names_mapping`; the files (`persons.txt`, `person.py`,
`persons.rs`) were inconsistent. Rename for clarity:

  rigour/names/person.py           → rigour/names/person_names.py
  rust/src/names/persons.rs        → rust/src/names/person_names.rs
  rust/data/names/persons.txt      → rust/data/names/person_names.txt
  tests/names/test_person.py       → tests/names/test_person_names.py
  rigour._core.persons_text()      → rigour._core.person_names_text()
  names::persons                   → names::person_names
  $OUT_DIR/persons.txt.zst         → $OUT_DIR/person_names.txt.zst

contrib/namesdb/Makefile's dump target writes to the new filename.
All consumer imports (rigour.names.__init__, regex_bench/bench.py,
test files, plans) updated.
