+++ docs/data.md
@@ -13,7 +13,7 @@ For name matching, we produce four classes of data assets:
     - *Compare normalisation* is designed to facilitate string-based comparison of company types, e.g. `Sp. z o.o.` becomes `spzoo`, `G.m.beschr. Haftung` becomes `gmbh`. 
     - *Generic normalisation* attempts a basic global alignment of organisation types. `Aktiengesellschaft` becomes `JSC`, `ОБЩЕСТВО С ОГРАНИЧЕННОЙ ОТВЕТСТВЕННОСТЬЮ` and  `Sp. z o.o.` become `LLC`.
 * **Organisation symbols** ([YAML](https://github.com/opensanctions/rigour/blob/main/resources/names/symbols.yml)) are mappings of common terms and phrases found in company names (like `Holding`, `Group`, `Company`, or `Technology`) across various spelling variants and languages (eg. `Holding`, `холдинг`, `集体`). Symbols do not imply a canonical form. Instead, they can be used as reciprocal synonyms and matched (and then eliminated) before string comparison is applied. This helps to avoid match biases introduced by long, but ultimately meaningless, name parts. Numbers in both their cardinal forms and ordinal forms are treated as a special type of symbol (eg. `12th`, `Nr. 12`, `№ 12`, XII). They can often be used to disambiguate two entities.
-* **Person names** ([TXT](https://github.com/opensanctions/rigour/blob/main/rigour/data/names/persons.txt)) is a database of hundreds of thousands of human name part mappings. Again, the idea is not to provide canonicalisation, but to be able to identify synonyms inside of a matching process (eg. `salman`, `салман`, `салмон`, `סלמאן`, `सलमान`, `サルマーン`, `살만`). This allows for name matching across scripts without the need to rely on machine-based transliteration. 
+* **Person names** ([TXT](https://github.com/opensanctions/rigour/blob/main/rust/data/names/person_names.txt)) is a database of hundreds of thousands of human name part mappings. Again, the idea is not to provide canonicalisation, but to be able to identify synonyms inside of a matching process (eg. `salman`, `салман`, `салмон`, `סלמאן`, `सलमान`, `サルマーン`, `살만`). This allows for name matching across scripts without the need to rely on machine-based transliteration. 
 
 ## Territories
