+++ plans/rust-tagger.md
@@ -41,7 +41,7 @@ becomes Rust-only?
 | `resources/text/ordinals.yml` | Python (`rigour/addresses/normalize.py`) and Rust (tagger build) | **Yes** for the Python side; Rust tagger `include_str!`s the same JSON internally |
 | `resources/names/symbols.yml` — `org_symbols`, `org_domains`, `person_symbols`, `person_nick`, `person_name_parts` | Rust-only — only the tagger consumes these | **No** |
 | Territory name aliases (stripped subset from `resources/territories/*.yml`) | Rust-only — tagger LOCATION symbols | **No** |
-| `persons.txt` (committed plain; `build.rs` compresses to OUT_DIR) | Rust-only — person tagger | **No** |
+| `person_names.txt` (committed plain; `build.rs` compresses to OUT_DIR) | Rust-only — person tagger | **No** |
 
 The Rust-only artifacts still live under `rust/data/…`, still get
 regenerated by `make rust-data`, still get covered by the CI no-diff
@@ -60,7 +60,7 @@ No `.pyi` stub, no Python-side accessor, no export from
 | 4 | `text/ordinals.yml` → `rust/data/text/ordinals.json` | `ordinals()` for the addresses/normalize consumer | `rigour/addresses/normalize.py` reads from Rust; `rigour/data/text/ordinals.py` deleted |
 | 5 | `symbols.yml` → `rust/data/names/symbols.json` | **None** — Rust-only artifact | Pure `make rust-data` diff; nothing observable until step 8 consumes it |
 | 6 | Territory subset → `rust/data/territory_names.jsonl`. Added as one extra emit in `generate_territories.py`; the full JSONL that `rigour.territories.*` consumes stays unchanged. | **None** | Same — pure data diff until step 8 |
-| 7 | `rust/data/names/persons.txt` (committed as plain UTF-8 from the namesdb pipeline). `rust/build.rs` zstd-compresses at build time into `OUT_DIR`; `names::persons` decodes on first access via `include_bytes!`. Plain text stays in git for diffability; the wheel ships ~2.7 MB of compressed bytes instead of 8.5 MB. | **None** | `rust/data/names/persons.txt` lands + `names::persons::raw()` loader; nothing consumes it yet until step 8 |
+| 7 | `rust/data/names/person_names.txt` (committed as plain UTF-8 from the namesdb pipeline). `rust/build.rs` zstd-compresses at build time into `OUT_DIR`; `names::person_names` decodes on first access via `include_bytes!`. Plain text stays in git for diffability; the wheel ships ~2.7 MB of compressed bytes instead of 8.5 MB. | **None** | `rust/data/names/person_names.txt` lands + `names::person_names::raw()` loader; nothing consumes it yet until step 8 |
 | 8 | Tagger port itself — `rust/src/names/tagger.rs`. Consumes steps 4/5/6/7 via `include_str!` / `include_bytes!`, plus the already-landed `org_types.json` for ORG_CLASS symbols. `Needles<Vec<Symbol>>` payload. Flag-keyed cache keyed on `(TaggerKind, Normalize, Cleanup)`. Python `rigour/names/tagging.py` becomes a thin wrapper. | `tag_org_name(name, normalize_flags, cleanup)`, `tag_person_name(name, normalize_flags, cleanup, infer_initials)` | Single-FFI tagger; `ahocorasick-rs` drops from `pyproject.toml`; nomenklatura's hot loop gets the throughput win |
 | 9 | Downstream adapters update: nomenklatura, yente, FTM switch from `normalizer=` to `normalize_flags=` on the tagger entry points. Driven by the corresponding PR in each repo. | n/a | Downstream tests pass against the new rigour |
 
@@ -133,7 +133,7 @@ fn build_tagger(kind: TaggerKind, flags: Normalize, cleanup: Cleanup) -> Tagger 
         }
         TaggerKind::Person => {
             // Load person_symbols, person_name_parts, person_nick,
-            // persons corpus via names::persons::raw().
+            // persons corpus via names::person_names::raw().
         }
     }
 
@@ -170,14 +170,14 @@ pub fn tag_org_name(
 fn load_ordinals() -> &'static [OrdinalSpec] { /* LazyLock<Vec<OrdinalSpec>> */ }
 fn load_symbols() -> &'static NameSymbols { /* same for rust/data/names/symbols.json */ }
 fn load_territory_names() -> &'static [TerritoryName] { /* for rust/data/territory_names.jsonl */ }
-fn load_person_corpus() -> &'static [PersonEntry] { /* parses names::persons::raw() lazily — the raw corpus is zstd-decoded once on first access, already handled by that module */ }
+fn load_person_corpus() -> &'static [PersonEntry] { /* parses names::person_names::raw() lazily — the raw corpus is zstd-decoded once on first access, already handled by that module */ }
 ```
 
 Each is a `LazyLock` that pays its load cost once per process. The
 persons corpus is the big one — ~8.5 MB of plain UTF-8 committed at
-`rust/data/names/persons.txt`, compressed to ~2.7 MB by `build.rs`
+`rust/data/names/person_names.txt`, compressed to ~2.7 MB by `build.rs`
 into `OUT_DIR` at crate-build time, and decompressed once on first
-access via `names::persons::raw()`. The tagger build then pays
+access via `names::person_names::raw()`. The tagger build then pays
 another ~50–200 ms for parsing + AC construction on top.
 
 ## Symbol payload: `Vec<Symbol>` vs `SmallVec<[Symbol; N]>`
