+++ rigour/_core.pyi
@@ -26,7 +26,7 @@ def obj_name_prefixes_list() -> list[str]: ...
 def name_split_phrases_list() -> list[str]: ...
 def generic_person_names_list() -> list[str]: ...
 def ordinals_dict() -> dict[int, list[str]]: ...
-def persons_text() -> str: ...
+def person_names_text() -> str: ...
 
 
 class SymbolCategory:
