+++ rust/Cargo.toml
@@ -59,9 +59,9 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 
 # zstd — compile-time compression of the person-names corpus
-# (`rust/data/names/persons.txt`, ~8.5 MB of text) into ~2–3 MB of
+# (`rust/data/names/person_names.txt`, ~8.5 MB of text) into ~2–3 MB of
 # bytes baked into the binary via OUT_DIR. Runtime decodes on first
-# access in `names::persons`. See `build.rs` for the compression step
+# access in `names::person_names`. See `build.rs` for the compression step
 # and `rust-tagger.md` step 7 for the design. Listed in both
 # [dependencies] (runtime decode) and [build-dependencies] (build.rs
 # uses the same crate for encoding).
