+++ rust/src/lib.rs
@@ -201,13 +201,13 @@ fn py_ordinals_dict() -> std::collections::HashMap<u32, Vec<String>> {
 // allocates a fresh `PyString` (the corpus is ~8.5 MB of text), so
 // Python consumers read it once and iterate locally. Will be retired
 // when the tagger ports to Rust (step 8 of plans/rust-tagger.md) —
-// at which point the Rust side parses `names::persons::raw()`
+// at which point the Rust side parses `names::person_names::raw()`
 // directly without crossing the FFI.
 #[cfg(feature = "python")]
 #[pyfunction]
-#[pyo3(name = "persons_text")]
-fn py_persons_text() -> &'static str {
-    names::persons::raw()
+#[pyo3(name = "person_names_text")]
+fn py_person_names_text() -> &'static str {
+    names::person_names::raw()
 }
 
 #[cfg(feature = "python")]
@@ -234,7 +234,7 @@ fn _core(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(py_name_split_phrases_list, m)?)?;
     m.add_function(wrap_pyfunction!(py_generic_person_names_list, m)?)?;
     m.add_function(wrap_pyfunction!(py_ordinals_dict, m)?)?;
-    m.add_function(wrap_pyfunction!(py_persons_text, m)?)?;
+    m.add_function(wrap_pyfunction!(py_person_names_text, m)?)?;
     m.add_class::<names::symbol::Symbol>()?;
     m.add_class::<names::symbol::SymbolCategory>()?;
     Ok(())
